# frontend/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Sprint analytics for Ultimate Frisbee trainings. Times inside a session are integer seconds `t` from its first sample; units are m, s, m/s, m/s² and bpm. Acceleration values are derived from 1 Hz GPS speed.",
    "title" : "Ultimate Analytics API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/sessions/{id}/segments" : {
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "overwrite",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SegmentCreate"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SegmentDto"
                }
              }
            },
            "description" : "Created"
          }
        },
        "summary" : "Create a segment; 409 if it overlaps another one, unless overwrite=true: then the overlapped segments are trimmed, split around it or removed",
        "tags" : [ "Segments" ]
      }
    },
    "/api/sessions/{id}/segments/merge" : {
      "post" : {
        "operationId" : "merge",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SegmentMerge"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SegmentDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Merge consecutive segments; the first keeps its drill type and label",
        "tags" : [ "Segments" ]
      }
    },
    "/api/sessions/{id}/segments/reset-from-laps" : {
      "post" : {
        "operationId" : "resetFromLaps",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "confirm",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SegmentDto"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Replace all segments with the lap-based ones; requires confirm=true",
        "tags" : [ "Segments" ]
      }
    },
    "/api/sessions/{id}/segments/{segmentId}" : {
      "delete" : {
        "operationId" : "delete_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "segmentId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "Segments" ]
      },
      "patch" : {
        "operationId" : "update_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "segmentId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SegmentPatch"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SegmentDto"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Edit bounds, drill type or label; 409 if the new bounds overlap another segment",
        "tags" : [ "Segments" ]
      }
    },
...
    "/api/sessions/{id}/segments/{segmentId}/split" : {
      "post" : {
        "operationId" : "split",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "path",
          "name" : "segmentId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SegmentSplit"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SegmentDto"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Split a segment at atT; both parts keep its drill type and label",
        "tags" : [ "Segments" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "SegmentCreate" : {
        "type" : "object",
        "properties" : {
          "drillTypeId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "endT" : {
            "type" : "integer",
            "format" : "int32"
          },
          "label" : {
            "type" : [ "string", "null" ]
          },
          "startT" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "endT", "startT" ]
      },
      "SegmentDto" : {
        "type" : "object",
        "properties" : {
          "drillTypeId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "endT" : {
            "type" : "integer",
            "format" : "int32"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "label" : {
            "type" : [ "string", "null" ]
          },
          "source" : {
            "type" : "string",
            "enum" : [ "LAP", "MANUAL" ]
          },
          "startT" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "endT", "id", "source", "startT" ]
      },
      "SegmentMerge" : {
        "type" : "object",
        "properties" : {
          "segmentIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        },
        "required" : [ "segmentIds" ]
      },
      "SegmentPatch" : {
        "type" : "object",
        "properties" : {
          "drillTypeId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "endT" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "label" : {
            "type" : [ "string", "null" ]
          },
          "startT" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          }
        }
      },
      "SegmentSplit" : {
        "type" : "object",
        "properties" : {
          "atT" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "atT" ]
      },
...
    }
  }
}